NEW: CONTENSIS-13034 - refactoring
