+++ bundle-es2015/events/event-operations.d.ts
@@ -1,9 +1,9 @@
 import { IHttpClient } from 'contensis-core-api';
-import { ContensisClient, EventsSubscription, IEventOperations } from '../models';
+import { ContensisClient, EventSubscription, IEventOperations } from '../models';
 export declare class EventOperations implements IEventOperations {
     private httpClient;
     private contensisClient;
     constructor(httpClient: IHttpClient, contensisClient: ContensisClient);
-    connectToEventsStream(eventsSubscription: EventsSubscription, eventsCallback: (eventsJSON: string, subscriptionId: string) => void): Promise<EventSource>;
+    connectToEventsStream(eventSubscription: EventSubscription, eventCallback: (message: any, subscriptionId: string) => void): Promise<EventSource>;
     private ensureIsBrowser;
 }
