+++ lib/events/event-operations.js
@@ -7,16 +7,16 @@ class EventOperations {
         this.httpClient = httpClient;
         this.contensisClient = contensisClient;
     }
-    connectToEventsStream(eventsSubscription, eventsCallback) {
+    connectToEventsStream(eventSubscription, eventCallback) {
         this.ensureIsBrowser('connectToEventsStream');
-        if (!eventsSubscription) {
-            throw new Error('A valid events subscription needs to be specified.');
+        if (!eventSubscription) {
+            throw new Error('A valid event subscription needs to be specified.');
         }
-        if (!eventsSubscription.topics || eventsSubscription.topics.length === 0) {
-            throw new Error('Valid events subscription topics need to be specified.');
+        if (!eventSubscription.topics || eventSubscription.topics.length === 0) {
+            throw new Error('Valid event subscription topics need to be specified.');
         }
-        if (!eventsCallback) {
-            throw new Error('A valid events callback needs to be specified.');
+        if (!eventCallback) {
+            throw new Error('A valid event callback needs to be specified.');
         }
         let params = this.contensisClient.getParams();
         let url = contensis_core_api_1.UrlBuilder.create('/api/management/projects/:projectId/events/sse/stream')
@@ -25,7 +25,7 @@ class EventOperations {
         return this.contensisClient.ensureBearerToken().then(() => {
             EventSource = sse_1.SSE;
             let eventSource = new sse_1.SSE(`${params.rootUrl}${url}`, {
-                payload: JSON.stringify(eventsSubscription),
+                payload: JSON.stringify(eventSubscription),
                 method: 'POST',
                 headers: this.contensisClient.getHeaders()
             });
@@ -42,7 +42,7 @@ class EventOperations {
                     doCallback = true;
                 }
                 if (doCallback) {
-                    eventsCallback(data, subscriptionId);
+                    eventCallback(data, subscriptionId);
                 }
             });
             eventSource.stream();
