+++ lib/models/EventSubscription.d.ts
@@ -1,0 +1,7 @@
+import { EventSubscriptionTopic } from './EventSubscriptionTopic';
+export interface EventSubscription {
+    topics: EventSubscriptionTopic[];
+    templates?: {
+        [key: string]: string;
+    };
+}
