+++ lib/models/EventSubscription.js
