--- lib/models/EventsSubscription.d.ts
@@ -1,7 +1,0 @@
-import { EventsSubscriptionTopic } from './EventsSubscriptionTopic';
-export interface EventsSubscription {
-    topics: EventsSubscriptionTopic[];
-    templates?: {
-        [key: string]: string;
-    };
-}
