+++ src/models/EventSubscription.ts
@@ -1,0 +1,6 @@
+import { EventSubscriptionTopic } from './EventSubscriptionTopic';
+
+export interface EventSubscription {
+    topics: EventSubscriptionTopic[];
+    templates?: { [key: string]: string };
+}
