--- src/models/EventsSubscription.ts
@@ -1,6 +1,0 @@
-import { EventsSubscriptionTopic } from './EventsSubscriptionTopic';
-
-export interface EventsSubscription {
-    topics: EventsSubscriptionTopic[];
-    templates?: { [key: string]: string };
-}
