+++ src/models/index.ts
@@ -14,8 +14,8 @@ export * from './Entry';
 export * from './EntryGetOptions';
 export * from './EntryListOptions';
 export * from './EntrySys';
-export * from './EventsSubscription';
-export * from './EventsSubscriptionTopic';
+export * from './EventSubscription';
+export * from './EventSubscriptionTopic';
 export * from './Group';
 export * from './GroupListOptions';
 export * from './IComponentOperations';
